Add load_env function in cmake
Introduced a new function 'load_env' in freshly created 'loadEnv.cmake' file. This new function reads a file and for each line not starting with a comment, it replaces spaces with nothing, splits the line on '=', and sets an environment variable with the parsed values. This addition will improve environment setup convenience.