Implement MongoDB connection in main.cpp
The main.cpp file has been modified to establish a connection to MongoDB as part of the program initialization. The MongoDB client is set up and a 'ping' command is sent to the 'admin' database in order to test the connection. Any exceptions during this process are also properly handled.